requirements.txt
openai
pandas
tiktoken
scipy
python-dotenv
